[modbus, context] Init modbus_context_1
