+++ modbus/Cargo.toml
@@ -12,5 +12,5 @@ derivative = "2.1.1"
 thiserror = "1.0"
 
 [features]
-
+#default = ["tokio-modbus"]
 test = []
